refactor(GraphQL): break into separate files
