+++ imports/plugins/core/graphql/server/resolvers/util/getCollectionFromCursor.js
@@ -1,0 +1,5 @@
+export default function getCollectionFromCursor(cursor) {
+  const { db } = cursor.options;
+  const collectionName = cursor.ns.slice(db.databaseName.length + 1);
+  return db.collection(collectionName);
+}
